Without Selenium

+++ db/models.py
@@ -1,6 +1,7 @@
+import decimal
 from typing import Sequence
 
-from sqlalchemy import String, Integer, ARRAY, Enum, Column
+from sqlalchemy import String, ARRAY, Enum, Column, DECIMAL
 from sqlalchemy.orm import declarative_base, mapped_column, Mapped
 
 from config import GenderTypes
@@ -15,7 +16,7 @@ class Product(Base):
     brand: Mapped[str] = mapped_column(String, nullable=False)
     category: Mapped[str] = mapped_column(String, nullable=False)
     description: Mapped[str] = mapped_column(String, nullable=False)
-    price: Mapped[int] = mapped_column(Integer, nullable=False)
+    price: Mapped[decimal] = mapped_column(DECIMAL, nullable=False)
     colors: Mapped[Sequence] = mapped_column(ARRAY(String), nullable=False)
     sizes: Mapped[Sequence] = mapped_column(ARRAY(String), nullable=False)
     images: Mapped[Sequence] = mapped_column(ARRAY(String), nullable=False)
